Don't attempt to rmdir the tags root itself
Or its parent directory.

For illustration, assume the tags root is "/mcp/spfs/tags" and we're
deleting a tag called "foo".

First, "/mcp/spfs/tags/foo" is deleted. `filepath` is
"/mcp/spfs/tags/foo".

This starts with tags root, so parent is "/mcp/spfs/tags". An
`rmdir /mcp/spfs/tags` is attempted. This might succeed if there were
no other tags. `filepath` is now "/mcp/spfs/tags".

This starts with tags root, so parent is "/mcp/spfs". An
`rmdir /mcp/spfs` is attempted!

Break out of the loop once `filepath` is the tags root. Don't attempt to
delete the tags root itself, or its parent.

+++ crates/spfs/src/storage/fs/tag.rs
@@ -190,9 +190,13 @@ impl TagStorage for FSRepository {
         // the lock file needs to be removed if the directory has any hope of being empty
         drop(lock);
 
+        let tags_root = self.tags_root();
         let mut filepath = filepath.as_path();
-        while filepath.starts_with(self.tags_root()) {
+        while filepath.starts_with(&tags_root) {
             if let Some(parent) = filepath.parent() {
+                if parent == tags_root {
+                    break;
+                }
                 tracing::trace!(?parent, "seeing if parent needs removing");
                 match tokio::fs::remove_dir(&parent).await {
                     Ok(_) => {
